UI: test reorg cleanup 2 (PKI) (#26353)
* rename pki-workflow-helpers > pki-helpers

* consolidate pki selectors into pki-selectors
