+++ ui/tests/acceptance/pki/pki-engine-route-cleanup-test.js
@@ -13,11 +13,13 @@ import enablePage from 'vault/tests/pages/settings/mount-secret-backend';
 import { click, currentURL, fillIn, visit } from '@ember/test-helpers';
 import { runCmd } from 'vault/tests/helpers/commands';
 import { GENERAL } from 'vault/tests/helpers/general-selectors';
-import { PKI_CONFIGURE_CREATE } from 'vault/tests/helpers/components/pki/pki-configure-create-selectors';
-import { PKI_KEYS } from 'vault/tests/helpers/components/pki/page/pki-keys-selectors';
-import { PKI_ISSUER_DETAILS } from 'vault/tests/helpers/components/pki/pki-issuer-details-selectors';
-import { PKI_ROLE_DETAILS } from 'vault/tests/helpers/components/pki/page/pki-role-details-selectors';
-import { PKI_ISSUER_LIST } from 'vault/tests/helpers/components/pki/page/pki-issuer-list-selectors';
+import {
+  PKI_CONFIGURE_CREATE,
+  PKI_ISSUER_DETAILS,
+  PKI_ISSUER_LIST,
+  PKI_KEYS,
+  PKI_ROLE_DETAILS,
+} from 'vault/tests/helpers/pki/pki-selectors';
 
 const OVERVIEW_BREADCRUMB = '[data-test-breadcrumbs] li:nth-of-type(2) > a';
 /**
