+++ ui/tests/acceptance/pki/pki-engine-workflow-test.js
@@ -16,13 +16,15 @@ import { runCmd, tokenWithPolicyCmd } from 'vault/tests/helpers/commands';
 import { create } from 'ember-cli-page-object';
 import flashMessage from 'vault/tests/pages/components/flash-message';
 import { GENERAL } from 'vault/tests/helpers/general-selectors';
-import { CERTIFICATES, clearRecords } from 'vault/tests/helpers/pki/pki-workflow-helpers';
-import { PKI_KEYS } from 'vault/tests/helpers/components/pki/page/pki-keys-selectors';
-import { PKI_ISSUER_DETAILS } from 'vault/tests/helpers/components/pki/pki-issuer-details-selectors';
-import { PKI_CONFIGURE_CREATE } from 'vault/tests/helpers/components/pki/pki-configure-create-selectors';
-import { PKI_CONFIG_EDIT } from 'vault/tests/helpers/components/pki/page/pki-configuration-edit-selectors';
-import { PKI_ROLE_DETAILS } from 'vault/tests/helpers/components/pki/page/pki-role-details-selectors';
-import { PKI_ISSUER_LIST } from 'vault/tests/helpers/components/pki/page/pki-issuer-list-selectors';
+import { CERTIFICATES, clearRecords } from 'vault/tests/helpers/pki/pki-helpers';
+import {
+  PKI_CONFIGURE_CREATE,
+  PKI_CONFIG_EDIT,
+  PKI_ISSUER_DETAILS,
+  PKI_ISSUER_LIST,
+  PKI_KEYS,
+  PKI_ROLE_DETAILS,
+} from 'vault/tests/helpers/pki/pki-selectors';
 
 const flash = create(flashMessage);
 const { unsupportedPem } = CERTIFICATES;
