--- ui/tests/helpers/components/pki/overview-selectors.ts
@@ -1,27 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: BUSL-1.1
- */
-
-export const PKI_OVERVIEW = {
-  issuersCardTitle: '[data-test-overview-card-title="Issuers"]',
-  issuersCardSubtitle: '[data-test-overview-card-subtitle="Issuers"]',
-  issuersCardLink: '[data-test-overview-card-container="Issuers"] a',
-  issuersCardOverviewNum: '[data-test-overview-card-container="Issuers"] h2',
-  rolesCardTitle: '[data-test-overview-card-title="Roles"]',
-  rolesCardSubtitle: '[data-test-overview-card-subtitle="Roles"]',
-  rolesCardLink: '[data-test-overview-card-container="Roles"] a',
-  rolesCardOverviewNum: '[data-test-overview-card-container="Roles"] h2',
-  issueCertificate: '[data-test-overview-card-title="Issue certificate"]',
-  issueCertificateInput: '[data-test-issue-certificate-input]',
-  issueCertificatePowerSearch: '[data-test-issue-certificate-input] span',
-  issueCertificateButton: '[data-test-issue-certificate-button]',
-  viewCertificate: '[data-test-overview-card-title="View certificate"]',
-  viewCertificateInput: '[data-test-view-certificate-input]',
-  viewCertificatePowerSearch: '[data-test-view-certificate-input] span',
-  viewCertificateButton: '[data-test-view-certificate-button]',
-  viewIssuerInput: '[data-test-issue-issuer-input]',
-  viewIssuerPowerSearch: '[data-test-issue-issuer-input] span',
-  viewIssuerButton: '[data-test-view-issuer-button]',
-  firstPowerSelectOption: '[data-option-index="0"]',
-};
