--- ui/tests/helpers/components/pki/page/pki-keys-selectors.ts
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: BUSL-1.1
- */
-
-export const PKI_KEYS = {
-  // key index
-  importKey: '[data-test-pki-key-import]',
-  generateKey: '[data-test-pki-key-generate]',
-  keyId: '[data-test-key="id"]',
-  keyName: '[data-test-key="name"]',
-  popupMenuDetails: '[data-test-key-menu-link="details"]',
-  popupMenuEdit: '[data-test-key-menu-link="edit"]',
-  // key details
-  keyDeleteButton: '[data-test-pki-key-delete]',
-  downloadButton: '[data-test-download-button]',
-  keyEditLink: '[data-test-pki-key-edit]',
-  nextStepsAlert: '[data-test-pki-key-next-steps]',
-};
