--- ui/tests/helpers/components/pki/page/pki-role-details-selectors.ts
@@ -1,17 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: BUSL-1.1
- */
-
-export const PKI_ROLE_DETAILS = {
-  issuerLabel: '[data-test-row-label="Issuer"]',
-  noStoreValue: '[data-test-value-div="Store in storage backend"]',
-  keyUsageValue: '[data-test-value-div="Key usage"]',
-  extKeyUsageValue: '[data-test-value-div="Ext key usage"]',
-  customTtlValue: '[data-test-value-div="Issued certificates expire after"]',
-  deleteRoleButton: '[data-test-pki-role-delete]',
-  generateCertLink: '[data-test-pki-role-generate-cert]',
-  signCertLink: '[data-test-pki-role-sign-cert]',
-  editRoleLink: '[data-test-pki-role-edit-link]',
-  createRoleLink: '[data-test-pki-role-create-link]',
-};
