--- ui/tests/helpers/components/pki/pki-delete-all-issuers-selectors.ts
@@ -1,11 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: BUSL-1.1
- */
-
-export const PKI_DELETE_ALL_ISSUERS = {
-  issuerLink: '[data-test-delete-all-issuers-link]',
-  deleteAllIssuerModal: '#confirmation-modal',
-  deleteAllIssuerInput: '[data-test-confirmation-modal-input="Delete all issuers?"]',
-  deleteAllIssuerButton: '[data-test-confirm-button="Delete all issuers?"]',
-};
