--- ui/tests/helpers/components/pki/pki-key-form-selectors.ts
@@ -1,8 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: BUSL-1.1
- */
-
-export const PKI_KEY_FORM = {
-  validationError: '[data-test-pki-key-validation-error]',
-};
