--- ui/tests/helpers/components/pki/pki-not-valid-after-form-selectors.ts
@@ -1,15 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: BUSL-1.1
- */
-
-export const PKI_NOT_VALID_AFTER = {
-  radioTtl: '[data-test-radio-button="ttl"]',
-  radioTtlLabel: '[data-test-radio-label="ttl"]',
-  radioDate: '[data-test-radio-button="not_after"]',
-  radioDateLabel: '[data-test-radio-label="specificDate"]',
-  ttlForm: '[data-test-ttl-inputs]',
-  ttlTimeInput: '[data-test-ttl-value="TTL"]',
-  ttlUnitInput: '[data-test-select="ttl-unit"]',
-  dateInput: '[data-test-input="not_after"]',
-};
