--- ui/tests/helpers/components/pki/pki-role-generate-selectors.ts
@@ -1,11 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: BUSL-1.1
- */
-
-export const PKI_ROLE_GENERATE = {
-  form: '[data-test-pki-generate-cert-form]',
-  optionsToggle: '[data-test-toggle-group="Subject Alternative Name (SAN) Options"]',
-  downloadButton: '[data-test-pki-cert-download-button]',
-  revokeButton: '[data-test-pki-cert-revoke-button]',
-};
