+++ ui/tests/integration/components/pki/page/pki-key-list-test.js
@@ -9,9 +9,9 @@ import { click, render } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 import { setupEngine } from 'ember-engines/test-support';
 import { setupMirage } from 'ember-cli-mirage/test-support';
-import { PKI_KEYS } from 'vault/tests/helpers/components/pki/page/pki-keys-selectors';
 import { STANDARD_META } from 'vault/tests/helpers/pagination';
 import { GENERAL } from 'vault/tests/helpers/general-selectors';
+import { PKI_KEYS } from 'vault/tests/helpers/pki/pki-selectors';
 
 module('Integration | Component | pki key list page', function (hooks) {
   setupRenderingTest(hooks);
