+++ ui/tests/integration/components/pki/page/pki-overview-test.js
@@ -9,7 +9,7 @@ import { render } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 import { setupEngine } from 'ember-engines/test-support';
 import { setupMirage } from 'ember-cli-mirage/test-support';
-import { PKI_OVERVIEW } from 'vault/tests/helpers/components/pki/overview-selectors';
+import { PKI_OVERVIEW } from 'vault/tests/helpers/pki/pki-selectors';
 
 module('Integration | Component | Page::PkiOverview', function (hooks) {
   setupRenderingTest(hooks);
