+++ ui/tests/integration/components/pki/pki-tidy-form-test.js
@@ -9,7 +9,7 @@ import { click, render, fillIn } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 import { setupEngine } from 'ember-engines/test-support';
 import { setupMirage } from 'ember-cli-mirage/test-support';
-import { PKI_TIDY_FORM } from 'vault/tests/helpers/components/pki/page/pki-tidy-form-selectors';
+import { PKI_TIDY_FORM } from 'vault/tests/helpers/pki/pki-selectors';
 
 module('Integration | Component | pki tidy form', function (hooks) {
   setupRenderingTest(hooks);
