+++ ui/tests/unit/serializers/pki/action-test.js
@@ -5,7 +5,7 @@
 
 import { module, test } from 'qunit';
 import { setupTest } from 'vault/tests/helpers';
-import { CERTIFICATES } from 'vault/tests/helpers/pki/pki-workflow-helpers';
+import { CERTIFICATES } from 'vault/tests/helpers/pki/pki-helpers';
 
 const { rootPem } = CERTIFICATES;
